Added complete CRUD

+++ src/components/common/AdminOptions/AdminOptions.tsx
@@ -1,9 +1,9 @@
-import { View, Text, TouchableOpacity, Alert } from 'react-native'
+import { View, Text, TouchableOpacity, Alert, Platform } from 'react-native'
 import React from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { DefaultState } from '../../../store'
 import { DataResponse } from '../../../api/types/app'
-import { WHITE, RED } from '../../../utils/constants';
+import { WHITE, RED, PRIMARY, BLACK } from '../../../utils/constants';
 import { deletePhoneDispatchAction } from '../../../store/app/dispatchers';
 
 interface IAdminOptions {
@@ -11,9 +11,11 @@ interface IAdminOptions {
     navigate: (route: string, params?: object) => void
     goBack: () => void
   }
+  handleEdit: (value: boolean) => void
+  isEditing: boolean
 }
 
-const AdminOptions: React.FunctionComponent<IAdminOptions> = ({ navigation }) => {
+const AdminOptions: React.FunctionComponent<IAdminOptions> = ({ navigation, handleEdit, isEditing }) => {
 
   const dispatch = useDispatch()
 
@@ -25,37 +27,59 @@ const AdminOptions: React.FunctionComponent<IAdminOptions> = ({ navigation }) =>
     navigation.goBack()
   }
 
-  const openPrompt = () => (
-    Alert.prompt(
-      'Remove phone',
-      'Are you sure you want to remove this phone?',
-      [
-        {
-          text: 'Cancel',
-          style: 'cancel'
-        },
-        {
-          text: 'OK',
-          onPress: () => deletePhone()
-        }
-      ],
-      'default'
-    )
-  )
+  const openPrompt = () => {
+    if (Platform.OS === 'ios') {
+      return (
+        Alert.prompt(
+          'Remove phone',
+          'Are you sure you want to remove this phone?',
+          [
+            {
+              text: 'Cancel',
+              style: 'cancel'
+            },
+            {
+              text: 'OK',
+              onPress: () => deletePhone()
+            }
+          ],
+          'default'
+        )
+      )
+    } else {
+      return (
+        Alert.alert(
+          'Remove phone',
+          'Are you sure you want to remove this phone?',
+          [
+            {
+              text: 'Cancel',
+              style: 'cancel'
+            },
+            {
+              text: 'OK',
+              onPress: () => deletePhone()
+            }
+          ],
+          { cancelable: false }
+        )
+      )
+    }
+  }
 
   return (
     <View style={{ flexDirection: 'row', justifyContent: 'space-between', marginHorizontal: 20, marginBottom: 20 }}>
       <TouchableOpacity
-      onPress={() => navigation.navigate('EditPhone', { id: phoneDetails._id })}
-      style={{ backgroundColor: WHITE, padding: 10, margin: 10, borderRadius: 10, width: '40%' }}
+        onPress={() => handleEdit(!isEditing)}
+        style={{ backgroundColor: isEditing ? PRIMARY : WHITE, padding: 10, margin: 10, borderRadius: 10, width: '40%' }}
       >
-        <Text style={{ textAlign: 'center' }}>Edit Phone</Text>
+        <Text style={{ textAlign: 'center', color: BLACK, fontWeight: 'bold' }}>{isEditing ? 'Editing now' : 'Edit phone'}</Text>
         </TouchableOpacity>
         <TouchableOpacity
           onPress={openPrompt}
           style={{ backgroundColor: RED, padding: 10, margin: 10, borderRadius: 10, width: '40%' }}
         >
-          <Text style={{ textAlign: 'center' }}>Delete Phone</Text>
+          <Text style={{ textAlign: 'center', color: BLACK, fontWeight: 'bold' }}>Delete Phone</Text>
       </TouchableOpacity>
     </View>
   )
